+++ lib/tinderfridge/devices/bricklet_joystick_v2.rb
@@ -1,0 +1,15 @@
+module Tinkerforge
+
+  class BrickletJoystickV2
+
+    # Returns the device's state.
+    def state
+      super.merge(
+        'position' => get_position,
+        'pressed'  => is_pressed,
+      )
+    end
+
+  end
+
+end